Move RLMSyncSubscription into its own file.
